+++ fork-visualizer/src/app/oldGraph/layout.js
@@ -1,0 +1,17 @@
+import { Inter } from 'next/font/google'
+import '@/app/globals.css'
+
+const inter = Inter({ subsets: ['latin'] })
+
+export const metadata = {
+  title: 'Create Next App',
+  description: 'Generated by create next app',
+}
+
+export default function RootLayout({ children }) {
+  return (
+    <div>
+        {children}
+    </div>
+  )
+}
